Changed expense datatype to double

+++ Code/app/src/main/java/com/avengers/enterpriseexpensetracker/modal/response/CategoryWiseTotalResponse.kt
@@ -6,26 +6,26 @@ import com.google.gson.annotations.SerializedName
 
 data class CategoryWiseTotalResponse(
     @SerializedName("accomadationTotalExpense")
-    private var accommodationExpense: Float = 0f,
-    private var otherTotalExpense: Float = 0f,
-    private var travelTotalExpense: Float = 0f,
-    private var foodTotalExpense: Float = 0f,
+    private var accommodationExpense: Double = 0.0,
+    private var otherTotalExpense: Double = 0.0,
+    private var travelTotalExpense: Double = 0.0,
+    private var foodTotalExpense: Double = 0.0,
     private var emailId: String?) : ApiResponse(), Parcelable {
 
     constructor(parcel: Parcel) : this(
-            parcel.readFloat(),
-            parcel.readFloat(),
-            parcel.readFloat(),
-            parcel.readFloat(),
+            parcel.readDouble(),
+            parcel.readDouble(),
+            parcel.readDouble(),
+            parcel.readDouble(),
             parcel.readString()) {
     }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         super.writeToParcel(parcel, flags)
-        parcel.writeFloat(accommodationExpense)
-        parcel.writeFloat(otherTotalExpense)
-        parcel.writeFloat(travelTotalExpense)
-        parcel.writeFloat(foodTotalExpense)
+        parcel.writeDouble(accommodationExpense)
+        parcel.writeDouble(otherTotalExpense)
+        parcel.writeDouble(travelTotalExpense)
+        parcel.writeDouble(foodTotalExpense)
         parcel.writeString(emailId)
     }
